+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Schema/NormalTextureInfo.cs
@@ -17,14 +17,21 @@ namespace GLTFast.Schema
         /// This value is ignored if normalTexture is not specified.
         /// This value is linear.
         /// </summary>
-        [JsonPropertyName("scale")]
-        public float Scale { get; set; } = 1.0f;
+        [JsonIgnore]
+        public float Scale { get; set; } = 1f;
+
+        [JsonPropertyName("scale"), JsonInclude]
+        internal float? ScaleSerialized
+        {
+            get => Mathematics.ApproximatelyOne(Scale) ? null : Scale;
+            set => Scale = value ?? 1f;
+        }
 
         internal override void GltfSerialize(JsonWriter writer)
         {
             writer.AddObject();
             GltfSerializeTextureInfo(writer);
-            if (math.abs(Scale - 1f) > Constants.epsilon)
+            if (!Mathematics.ApproximatelyOne(Scale))
             {
                 writer.AddProperty("scale", Scale);
             }