+++ Packages/com.unity.cloud.gltfast/CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `Camera.Type` ⇒ `CameraType`
     - `Material.AlphaMode` ⇒ `AlphaMode`
     - `MeshGpuInstancing.Attributes` ⇒ `InstancesAttributes`
+    - Applied PascalCase on `MaterialIor.DefaultIndexOfRefraction`
   - Type changes
     - `float[]` ⇒ `List<float>`
       - `Node.Weights`